+++ Core/Inc/util.h
@@ -1,0 +1,16 @@
+/*
+ * util.h
+ *
+ *  Created on: 15 de ago de 2021
+ *      Author: Felipe Telles
+ */
+
+#ifndef INC_UTIL_H_
+#define INC_UTIL_H_
+
+#include "main.h"
+#include "speed_calc.h"
+
+speed_pin_e get_speed_pin(uint16_t pin);
+
+#endif /* INC_UTIL_H_ */
